Move connections store to queries
